backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
apscheduler==3.10.4
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
pytz==2024.2
